Moved base class for running services to `common` dir
