+++ x-pack/plugins/security_solution/scripts/endpoint/agent_emulator/services/action_responder.ts
@@ -6,17 +6,17 @@
  */
 
 import { set } from 'lodash';
-import { Client } from '@elastic/elasticsearch';
-import { ToolingLog } from '@kbn/tooling-log';
-import { KbnClient } from '@kbn/test';
-import { BaseRunningService } from './base_running_service';
+import type { Client } from '@elastic/elasticsearch';
+import type { ToolingLog } from '@kbn/tooling-log';
+import type { KbnClient } from '@kbn/test';
+import { BaseRunningService } from '../../common/base_running_service';
 // FIXME:PT move these to a common directory or to `agent_emulator` once action reponder is removed
 import {
   fetchEndpointActionList,
   sendEndpointActionResponse,
   sendFleetActionResponse,
 } from '../../action_responder/utils';
-import { ActionDetails } from '../../../../common/endpoint/types';
+import type { ActionDetails } from '../../../../common/endpoint/types';
 
 /**
  * Base class for start/stopping background services
