+++ x-pack/plugins/security_solution/scripts/endpoint/agent_emulator/services/keep_alive.ts
@@ -10,7 +10,7 @@ import {
   fetchEndpointMetadataList,
   sendEndpointMetadataUpdate,
 } from '../../common/endpoint_metadata_services';
-import { BaseRunningService } from './base_running_service';
+import { BaseRunningService } from '../../common/base_running_service';
 
 export class AgentKeepAliveService extends BaseRunningService {
   protected async run(): Promise<void> {
